# frontend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: TRCustoms
  version: 0.0.0
  description: An API for TRCustoms.org
paths:
...
  /api/levels/:
    get:
      operationId: levels_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - name: sort
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      tags:
      - levels
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLevelListingList'
          description: ''
    post:
      operationId: levels_create
      tags:
      - levels
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LevelDetails'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LevelDetails'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LevelDetails'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelDetails'
          description: ''
...
components:
  schemas:
...
    EngineNested:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
      required:
      - id
      - name
...
    GenreNested:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
      required:
      - id
      - name
...
    LevelDetails:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 5000
        engine:
          allOf:
          - $ref: '#/components/schemas/EngineNested'
          readOnly: true
        engine_id:
          type: integer
          writeOnly: true
        duration:
          allOf:
          - $ref: '#/components/schemas/LevelDurationNested'
          readOnly: true
        duration_id:
          type: integer
          writeOnly: true
        difficulty:
          allOf:
          - $ref: '#/components/schemas/LevelDifficultyNested'
          readOnly: true
        difficulty_id:
          type: integer
          writeOnly: true
        genres:
          type: array
          items:
            $ref: '#/components/schemas/GenreNested'
          readOnly: true
        genre_ids:
          type: array
          items:
            type: integer
            writeOnly: true
          writeOnly: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagNested'
          readOnly: true
        tag_ids:
          type: array
          items:
            type: integer
            writeOnly: true
          writeOnly: true
        authors:
          type: array
          items:
            $ref: '#/components/schemas/UserNested'
          readOnly: true
        author_ids:
          type: array
          items:
            type: integer
            writeOnly: true
          writeOnly: true
        uploader:
          allOf:
          - $ref: '#/components/schemas/UserNested'
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        cover:
          allOf:
          - $ref: '#/components/schemas/UploadedFileNested'
          readOnly: true
        screenshots:
          type: array
          items:
            $ref: '#/components/schemas/LevelScreenshot'
          readOnly: true
        external_links:
          type: array
          items:
            $ref: '#/components/schemas/LevelExternalLink'
        last_updated:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        last_user_content_updated:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        last_file:
          allOf:
          - $ref: '#/components/schemas/LevelFile'
          readOnly: true
        download_count:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          readOnly: true
        rating_count:
          type: integer
          maximum: 2147483647
          minimum: 0
          readOnly: true
        review_count:
          type: integer
          maximum: 2147483647
          minimum: 0
          readOnly: true
        walkthrough_count:
          type: integer
          maximum: 2147483647
          minimum: 0
          readOnly: true
        is_approved:
          type: boolean
          readOnly: true
        rejection_reason:
          type: string
          nullable: true
          readOnly: true
        rating_class:
          allOf:
          - $ref: '#/components/schemas/RatingClassNested'
          readOnly: true
        cover_id:
          type: integer
          writeOnly: true
        screenshot_ids:
          type: array
          items:
            type: integer
            writeOnly: true
          writeOnly: true
        files:
          type: array
          items:
            $ref: '#/components/schemas/LevelFile'
          readOnly: true
        file_id:
          type: integer
          writeOnly: true
        trle_id:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
      required:
      - author_ids
      - authors
      - cover
      - cover_id
      - created
      - description
      - difficulty
      - difficulty_id
      - download_count
      - duration
      - duration_id
      - engine
      - engine_id
      - file_id
      - files
      - genre_ids
      - genres
      - id
      - is_approved
      - last_file
      - last_updated
      - last_user_content_updated
      - name
      - rating_class
      - rating_count
      - rejection_reason
      - review_count
      - screenshot_ids
      - screenshots
      - tag_ids
      - tags
      - uploader
      - walkthrough_count
...
    LevelDifficultyNested:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
      required:
      - id
      - name
...
    LevelDurationNested:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
      required:
      - id
      - name
    LevelExternalLink:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        position:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        url:
          type: string
          format: uri
          maxLength: 256
        link_type:
          $ref: '#/components/schemas/LinkTypeEnum'
      required:
      - id
      - link_type
      - position
      - url
    LevelFile:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        version:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        size:
          type: integer
          nullable: true
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        url:
          type: string
          nullable: true
          readOnly: true
      required:
      - created
      - id
      - size
      - url
      - version
    LevelListing:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        description:
          type: string
          nullable: true
          maxLength: 5000
        engine:
          allOf:
          - $ref: '#/components/schemas/EngineNested'
          readOnly: true
        engine_id:
          type: integer
          writeOnly: true
        duration:
          allOf:
          - $ref: '#/components/schemas/LevelDurationNested'
          readOnly: true
        duration_id:
          type: integer
          writeOnly: true
        difficulty:
          allOf:
          - $ref: '#/components/schemas/LevelDifficultyNested'
          readOnly: true
        difficulty_id:
          type: integer
          writeOnly: true
        genres:
          type: array
          items:
            $ref: '#/components/schemas/GenreNested'
          readOnly: true
        genre_ids:
          type: array
          items:
            type: integer
            writeOnly: true
          writeOnly: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagNested'
          readOnly: true
        tag_ids:
          type: array
          items:
            type: integer
            writeOnly: true
          writeOnly: true
        authors:
          type: array
          items:
            $ref: '#/components/schemas/UserNested'
          readOnly: true
        author_ids:
          type: array
          items:
            type: integer
            writeOnly: true
          writeOnly: true
        uploader:
          allOf:
          - $ref: '#/components/schemas/UserNested'
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        cover:
          allOf:
          - $ref: '#/components/schemas/UploadedFileNested'
          readOnly: true
        screenshots:
          type: array
          items:
            $ref: '#/components/schemas/LevelScreenshot'
          readOnly: true
        external_links:
          type: array
          items:
            $ref: '#/components/schemas/LevelExternalLink'
        last_updated:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        last_user_content_updated:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        last_file:
          allOf:
          - $ref: '#/components/schemas/LevelFile'
          readOnly: true
        download_count:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          readOnly: true
        rating_count:
          type: integer
          maximum: 2147483647
          minimum: 0
          readOnly: true
        review_count:
          type: integer
          maximum: 2147483647
          minimum: 0
          readOnly: true
        walkthrough_count:
          type: integer
          maximum: 2147483647
          minimum: 0
          readOnly: true
        is_approved:
          type: boolean
          readOnly: true
        rejection_reason:
          type: string
          nullable: true
          readOnly: true
        rating_class:
          allOf:
          - $ref: '#/components/schemas/RatingClassNested'
          readOnly: true
      required:
      - author_ids
      - authors
      - cover
      - created
      - difficulty
      - difficulty_id
      - download_count
      - duration
      - duration_id
      - engine
      - engine_id
      - genre_ids
      - genres
      - id
      - is_approved
      - last_file
      - last_updated
      - last_user_content_updated
      - name
      - rating_class
      - rating_count
      - rejection_reason
      - review_count
      - screenshots
      - tag_ids
      - tags
      - uploader
      - walkthrough_count
...
    LevelScreenshot:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        position:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        file:
          allOf:
          - $ref: '#/components/schemas/UploadedFileNested'
          readOnly: true
      required:
      - file
      - id
    LinkTypeEnum:
      enum:
      - sh
      - ma
      type: string
      description: |-
        * `sh` - Showcase
        * `ma` - Main
...
    PaginatedLevelListingList:
      type: object
      required:
      - current_page
      - last_page
      - total_count
      - items_on_page
      - next
      - previous
      - results
      - disable_paging
      properties:
        current_page:
          type: integer
        last_page:
          type: integer
        total_count:
          type: integer
        items_on_page:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/LevelListing'
        disable_paging:
          type: boolean
...
    RatingClassNested:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 30
        position:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
      required:
      - id
      - name
...
    TagNested:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
      required:
      - id
      - name
...
    UploadTypeEnum:
      enum:
      - up
      - lb
      - ls
      - lf
      - at
      - ec
      type: string
      description: |-
        * `up` - User picture
        * `lb` - Level cover image
        * `ls` - Level screenshot
        * `lf` - Level file
        * `at` - Post attachment
        * `ec` - Event cover image
...
    UploadedFileNested:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        content:
          type: string
          format: uri
          nullable: true
        upload_type:
          $ref: '#/components/schemas/UploadTypeEnum'
        url:
          type: string
          nullable: true
          readOnly: true
        size:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          readOnly: true
        md5sum:
          type: string
          nullable: true
          readOnly: true
      required:
      - id
      - md5sum
      - size
      - upload_type
      - url
...
    UserNested:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        picture:
          allOf:
          - $ref: '#/components/schemas/UploadedFileNested'
          readOnly: true
      required:
      - id
      - picture
      - username
...